dashboard/requirements.txt
# Correct content for ./dashboard/requirements.txt

streamlit
pandas
google-generativeai
watchdog